Move PoliciesPool into policy package
